+++ front/src/components/Profile/friend_request.tsx
@@ -5,90 +5,14 @@ import { socket } from "../../App";
 import { useNavigate, useParams } from "react-router-dom";
 import { useAppSelector } from '../../redux/Hook';
 
-export function InviteButton(props: { user: any }) {
-    const { user } = props;
-
-    const pathname = window.location.pathname;
-    const pathArray = pathname.split('/');
-    // const id = pathArray[pathArray.length - 1];
-    const [ReqStatus, setStatus] = useState<string>('+ Add Friend');
-    const [myVar, setMyVar] = useState<boolean>(true);
-    let { id } = useParams();
-    const myToken = useAppSelector(state => state.user.myToken);
 
 
-    const ifFriend = async () => {
-        console.log("ca rentre ici");
-        await fetch(`${process.env.REACT_APP_BACK}user/friend/check`, {
-            method: 'POST',
-            body: JSON.stringify({
-                myId: user.id,
-                friendId: id,
-            }),
-            headers: {
-                'Content-Type': 'application/json',
-                'Authorization': `Bearer ${myToken}`,
-            },
-        })
-            .then(async response => {
-                console.log("truefalse")
-                const data = await response.json();
-                console.log(data);
-                // check for error response
-                if (response.ok) {
-                    setMyVar(!data);
-                }
-
-            })
-    }
-
-    const sendFriendRequest = async () => {
-        await new Promise((resolve) => setTimeout(resolve, 1000));
-
-        await fetch(`${process.env.REACT_APP_BACK}user/friend-request/send/${id}`, {
-            method: 'POST',
-            body: JSON.stringify({ userId: user.id }),
-            headers: {
-                'Content-Type': 'application/json',
-                'Authorization': `Bearer ${myToken}`,
-
-            },
-        });
-        setStatus('Pending');
-    }
-
-    useEffect(() => {
-
-        const checkFriendRequest = async () => {
-            const response = await fetch(`${process.env.REACT_APP_BACK}user/friend-request/status/${id}`, {
-                method: 'POST',
-                body: JSON.stringify({ userId: user.id }),
-                headers: {
-                    'Content-Type': 'application/json',
-                    'Authorization': `Bearer ${myToken}`,
-             },
-            });
-            const data = await response.json()
-            if (data.ReqStatus)
-                setStatus(data.ReqStatus);
-            else
-                setStatus("+ Add Friend");
-        }
-        console.log("friend id", id);
-        if (id) {
-            ifFriend();
-            checkFriendRequest();
-        }
-    }, [id, user]);
 
 
-    return (myVar) ? (
 
-        <button className="reqButton pointer white width_50" onClick={_ => (setMyVar(false), sendFriendRequest())} >
-            {ReqStatus}
-        </button>
-    ) : <></>;
-}
+
+
+
 
 
 export function Friends(props: { user: IUser }) {
